+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### 0.10.5
 - :arrow_up: Changed fake notifications period from 24 hours to 4 days
+- :arrow_up: Added owner mkey to mobile notification payload
 
 ### 0.10.4
 - :hammer: Fixed bug with cron workers wrapper